task03: completed

+++ domain/src/main/java/com/ivos/domain/entities/Weather.kt
@@ -1,9 +1,9 @@
 package com.ivos.domain.entities
 
-import android.icu.util.Calendar
+import java.util.Calendar
 
 data class Weather(
-    val tempC: Float,
+    val tempC: Double,
     val description: String,
     val iconUrl: String,
     val date: Calendar,
